Add support for multiple song owners
Each song has now a “owners“ array field (instead of the “owner“ as
single value string. Meteor 'pushSong' and 'clear' use Mongo.update() to
update the song (if “owners“ has more than on value).
The required criteria for an update (instead of an insert) is an
identical “_id“. MongoDB's upsert doesnt seem to work when inserting
with a predfined "_id" field. Maybe in future versions.